+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - xxxx-xx-xx
+
+### Breaking Changes
+
+- Add retry mechanism for transport errors
+    - Add `use_retries` parameter to:
+        - `MCUmgrClient::shell_execute`
+        - `Connection::execute_raw_command`
+
+### Changes
+
+- Decrease default timeout to `500 ms`
+
 ## [0.10.0] - 2026-02-09
 
 ### Breaking Changes
@@ -56,7 +69,7 @@ For that reason, the project was renamed to `mcumgr-toolkit`:
 - Add CLI commands:
   - `firmware`
     - `update`
-- Increase default timeout to `500 ms`
+- Increase default timeout to `10000 ms`
 
 ### Fixes
 
@@ -290,6 +303,7 @@ Initial release, not feature complete yet.
 
 Primarily to test release workflow.
 
+[0.11.0]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.8.1...0.9.0
 [0.8.1]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.8.0...0.8.1